requirements.txt
discord.py==1.6.0
hentai==3.2.5
translate==3.5.0
PyNaCl==1.3.0
numpy==1.20.0
